Main Commit
This is the main commit, tremble at its almighty main commitness. Fear the main commit. Worship the main commit. In the face of the main commit, we are nothing. The commit is us now. It is our life. This is how things are. Forever, until the end of time, which is also the end of the commit, which will never happen because the main commit is also now the main timeline.