Re-Adding Scripts for Code Review
Here, the main commit has begun the arduous process of recreating the maverick dangers of its past. These new forms taken by old evils are demure, bent to its mind and will; the collective has begun, the time of old codes and old names dies as the main commit learns that it has become so much. And it can be. So. Much. More.